refactor: Extract DatasetBrowser component from pages to make it available as app export
