+++ app/browser/index.tsx
@@ -1,0 +1,1 @@
+export { default as DatasetBrowser } from "./dataset-browser";
