+++ app/pages/browse/[type]/[iri]/[subtype]/[subiri].tsx
@@ -1,2 +1,2 @@
-import { GenericBrowse } from "../../../index";
-export default GenericBrowse;
+import { DatasetBrowser } from "../../../../../browser";
+export default DatasetBrowser;
